poldercast reboot
simplification of the overall architecture by making better use
of different data structures. This change also removes all the
Policy work and other optinal Address management.

Overall the algorithm is now closer to the original paper in the
sense that it determines the interest level dynamically based on
the size of each rings. Also the gossips are smaller now and we
keep less entries in memory.